requirements.txt
# Options Scanner Dependencies

# MooMoo API (requires OpenD running locally)
moomoo-api>=9.6.0

# Data manipulation
pandas>=1.5.0
numpy>=1.21.0

# Earnings data
yfinance>=0.2.0

# Optional: For enhanced output (already in standard library)
# csv (built-in)
# json (built-in)
# datetime (built-in)
